Fix format of _size

+++ example/ex-list01.c
@@ -38,7 +38,7 @@ int main(void)
     list_string_get_str (tmpstr, grow, false);
 
     // print out
-    printf("Number of elements = %lu\n", (unsigned long) list_string_size(grow));
+    printf("Number of elements = %zu\n", list_string_size(grow));
     printf("Final string = %s\n", string_get_cstr(tmpstr));
 
     // Let's iterate over each element of the list
